drivers: timer: clean up tick clamps for the unsigned interface
With sys_clock_set_timeout() and sys_clock_announce() now taking an
unsigned tick count, the per-driver clamps and casts that assumed a
signed argument are no longer right:

- The (int32_t) cast on the CLAMP upper bound (MAX_TICKS and friends)
  is unnecessary now that ticks is unsigned; drop it.

- CLAMP(ticks - 1, 0, MAX) underflows when ticks == 0: the unsigned
  ticks - 1 wraps to UINT32_MAX and clamps to MAX_TICKS, turning an
  'announce ASAP' request into a maximum sleep. Use
  CLAMP(ticks, 1, MAX) - 1, which floors at 1 before subtracting so
  ticks 0 and 1 both map to 0 with no underflow. stm32_lptim keeps its
  minimum of one scheduled tick via CLAMP(ticks, 2, base) - 1. The
  renesas_rx and renesas_ra_ulpt MAX_TICKS can be 0 (a short counter
  relative to CYCLES_PER_TICK), which would invert that clamp range, so
  they keep the original guarded subtract-then-cap, to be revisited.

- CLAMP(ticks, 0, MAX) for unsigned ticks has a pointless lower bound;
  use MIN(ticks, MAX).

- Drop the now-redundant (int32_t) casts on the sys_clock_announce()
  arguments, and make stm32wb0's dticks unsigned to match.

The recurring 'ticks - 1' is most likely the sign of a suboptimal
computation, but that will be dealt with elsewhere.

No functional change beyond the ticks == 0 underflow fix.

+++ drivers/timer/ambiq_stimer.c
@@ -146,7 +146,7 @@ void sys_clock_set_timeout(uint32_t ticks, bool idle)
 	}
 
 	/* Adjust the ticks to the range of [1, MAX_TICKS]. */
-	ticks = CLAMP(ticks, 1, (int32_t)MAX_TICKS);
+	ticks = CLAMP(ticks, 1, MAX_TICKS);
 
 	k_spinlock_key_t key = k_spin_lock(&g_lock);
